hound::commands

Module messages

Message API

Commands used to send messages in hound.

Structs

Message 🔒

An individual message from a user.

Functions

list_messages

Lists all messages from the backend.

post_message

Submits a new message from the local user to the backend.

subscribe_to_messages

Subscribes to real time updates from the backend for new messages.