Enum hound::Error

pub enum Error {
    IoError(Error),
    FormatError(&'static str),
    TooWide,
    UnfinishedSample,
    Unsupported,
}

The error type for operations on WavReader and WavWriter.

Variants

IoError

An IO error occured in the underlying reader or writer.

FormatError

Ill-formed WAVE data was encountered.

TooWide

The sample has more bits than the destination type.

When iterating using the samples iterator, this means that the destination type (produced by the iterator) is not wide enough to hold the sample. When writing, this means that the sample cannot be written, because it requires more bits than the bits per sample specified.

UnfinishedSample

The number of samples written is not a multiple of the number of channels.

Unsupported

The format is not supported.