Function hound::read_wave_header

pub fn read_wave_header<R: Read>(reader: &mut R) -> Result<u64>

Reads the RIFF WAVE header, returns the supposed file size.

This function can be used to quickly check if the file could be a wav file by reading 12 bytes of the header. If an Ok is returned, the file is likely a wav file. If an Err is returned, it is definitely not a wav file.

The returned file size cannot be larger than 232 + 7 bytes.