Crate hound

Hound, a WAV encoding and decoding library.

Examples

The following example renders a 440 Hz sine wave, and stores it as as a mono wav file with a sample rate of 44.1 kHz and 16 bits per sample.

use std::f32::consts::PI;
use std::i16;
use hound;

let spec = hound::WavSpec {
    channels: 1,
    sample_rate: 44100,
    bits_per_sample: 16
};
let mut writer = hound::WavWriter::create("sine.wav", spec).unwrap();
for t in (0 .. 44100).map(|x| x as f32 / 44100.0) {
    let sample = (t * 440.0 * 2.0 * PI).sin();
    let amplitude = i16::MAX as f32;
    writer.write_sample((sample * amplitude) as i16).unwrap();
}
writer.finalize().unwrap();

The following example computes the RMS (root mean square) of an audio file.

use hound;

let mut reader = hound::WavReader::open("testsamples/pop.wav").unwrap();
let (sqr_sum, n) = reader.samples::<i16>()
                         .fold((0_f64, 0_u32), |(sqr_sum, n), s| {
    let sample = s.unwrap() as f64;
    (sqr_sum + sample * sample, n + 1)
});
println!("RMS is {}", (sqr_sum / n as f64).sqrt());

Structs

WavReader
A reader that reads the WAVE format from the underlying reader.

WavSamples
An iterator that yields samples of type S read from a WavReader.

WavSpec
Specifies properties of the audio data.

WavWriter
A writer that accepts samples and writes the WAVE format.

Enums

Error
The error type for operations on WavReader and WavWriter.

Traits

Sample
A type that can be used to represent audio samples.

Type Definitions

Result
A type for results generated by Hound where the error type is hard-wired.