Crate hound

Hound, a WAV encoding library.

TODO: Add some introductory text here.

The following example renders a 440 Hz sine wave, and stores it as as a mono wav file with a sample rate of 44.1 kHz and 16 bits per sample.

#![feature(core)]

use std::fs;
use std::f32::consts::PI;
use std::iter::IntoIterator;
use std::num;
use hound;

let spec = hound::WavSpec {
    channels: 1,
    sample_rate: 44100,
    bits_per_sample: 16
};
// TODO: ensure that the type can be inferred.
let writer_res = hound::WavWriter::<fs::File>::create("sine.wav", spec);
let mut writer = writer_res.ok().unwrap();
for t in (0 .. 44100).map(|x| x as f32 / 44100.0) {
    let sample = (t * 440.0 * 2.0 * PI).sin();
    let amplitude: i16 = num::Int::max_value();
    writer.write_sample((sample * amplitude as f32) as i16).ok().unwrap();
}
writer.finalize().ok().unwrap();

Structs

WavSpec
Specifies properties of the audio data.

WavWriter
A writer that accepts samples and writes the WAVE format.

Traits

Sample
A type that can be used to represent audio samples.