Enum hound::Error

pub enum Error {
    IoError(Error),
    FormatError(&'static str),
    TooWide,
    UnfinishedSample,
    Unsupported,
    InvalidSampleFormat,
}

The error type for operations on WavReader and WavWriter.

Variants
IoError(Error)

An IO error occured in the underlying reader or writer.

FormatError(&'static str)

Ill-formed WAVE data was encountered.

TooWide

The sample has more bits than the destination type.

When iterating using the samples iterator, this means that the destination type (produced by the iterator) is not wide enough to hold the sample. When writing, this means that the sample cannot be written, because it requires more bits than the bits per sample specified.

UnfinishedSample

The number of samples written is not a multiple of the number of channels.

Unsupported

The format is not supported.

InvalidSampleFormat

The sample format is different than the destination format.

When iterating using the samples iterator, this means the destination type (produced by the iterator) has a different sample format than the samples in the wav file.

For example, this will occur if the user attempts to produce i32 samples (which have a SampleFormat::Int) from a wav file that contains floating point data (SampleFormat::Float).