Struct hound::WavReader

pub struct WavReader<R> {
    // some fields omitted
}

A reader that reads the WAVE format from the underlying reader.

A WavReader is a streaming reader. It reads data from the underlying reader on demand, and it reads no more than strictly necessary. No internal buffering is performed on the underlying reader.

Methods

impl<R> WavReader<R> where R: Read

fn new(reader: R) -> Result<WavReader<R>>

Attempts to create a reader that reads the WAVE format.

The header is read immediately. Reading the data will be done on demand.

fn spec(&self) -> &WavSpec

Returns information about the WAVE file.

fn samples<'wr, S: Sample>(&'wr mut self) -> WavSamples<'wr, R, S>

Returns an iterator over all samples.

The channel data is is interleaved. The iterator is streaming. That is, if you call this method once, read a few samples, and call this method again, the second iterator will not start again from the beginning of the file, it will continue where the first iterator stopped.

fn duration(&self) -> u32

Returns the duration of the file in samples.

The duration is independent of the number of channels. It is expressed in units of samples. The duration in seconds can be obtained by dividing this number by the sample rate. The duration is independent of how many samples have been read already.

fn len(&self) -> u32

Returns the number of values that the sample iterator will yield.

The length of the file is its duration (in samples) times the number of channels. The length is independent of how many samples have been read already.

impl WavReader<BufReader<File>>

fn open<P: AsRef<Path>>(filename: P) -> Result<WavReader<BufReader<File>>>

Attempts to create a reader that reads from the specified file.

This is a convenience constructor that opens a File, wraps it in a BufReader and then constructs a WavReader from it.